backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.4.1",
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.5",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.2",
    "jq>=1.11.0",
    "langchain>=1.2.13",
    "langchain-astradb>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-nvidia-ai-endpoints>=1.2.1",
    "langchain-openai>=1.1.12",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.3",
    "langgraph-cli[inmem]>=0.4.19",
    "langgraph-supervisor>=0.0.31",
    "matplotlib>=3.10.8",
    "neo4j==5.14",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdfminer>=20191125",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.9.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "tiktoken>=0.12.0",
    "unstructured>=0.21.5",
    "uvicorn>=0.42.0",
]
